Revert "feat: custom user agent Geo changes for UI handoff (#11632)"
This reverts commit 01bfa8f692737bd14422f7dc2eae11ed00c19048.
