+++ packages/api-rest/package.json
@@ -56,7 +56,7 @@
       "name": "API (rest client)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, RestAPI }",
-      "limit": "31.5 kB"
+      "limit": "31 kB"
     }
   ],
   "jest": {
