+++ packages/auth/package.json
@@ -60,7 +60,7 @@
       "name": "Auth (top-level class)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, Auth }",
-      "limit": "55.15 kB"
+      "limit": "55.1 kB"
     }
   ],
   "jest": {
