--- packages/geo/src/internals/utils.ts
@@ -1,15 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { Category, CustomUserAgentDetails, GeoAction } from '@aws-amplify/core';
-
-export const getGeoUserAgentDetails = (
-	action: GeoAction,
-	customUserAgentDetails?: CustomUserAgentDetails
-): CustomUserAgentDetails => {
-	return {
-		category: Category.Geo,
-		action,
-		...customUserAgentDetails,
-	};
-};
