+++ packages/geo/src/types/Provider.ts
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { CustomUserAgentDetails } from '@aws-amplify/core';
 import {
 	SearchByTextOptions,
 	SearchByCoordinatesOptions,
@@ -36,55 +35,42 @@ export interface GeoProvider {
 	getDefaultMap(): MapStyle;
 
 	// search by a text string and return a list of places
-	searchByText(
-		text: string,
-		options?: SearchByTextOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
-	): Promise<Place[]>;
+	searchByText(text: string, options?: SearchByTextOptions): Promise<Place[]>;
 
 	// search by coordinates and return a matching place
 	searchByCoordinates(
 		coordinates: Coordinates,
-		options?: SearchByCoordinatesOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: SearchByCoordinatesOptions
 	): Promise<Place>;
 
 	searchForSuggestions(
 		text: string,
-		options?: SearchByTextOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: SearchByTextOptions
 	): Promise<SearchForSuggestionsResults>;
 
 	searchByPlaceId(
 		placeId: string,
-		options?: searchByPlaceIdOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: searchByPlaceIdOptions
 	): Promise<Place | undefined>;
 
 	// create geofences
 	saveGeofences(
 		geofences: GeofenceInput[],
-		options?: GeofenceOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: GeofenceOptions
 	): Promise<SaveGeofencesResults>;
 
 	// get a single geofence
 	getGeofence(
 		geofenceId: GeofenceId,
-		options?: ListGeofenceOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: ListGeofenceOptions
 	): Promise<Geofence>;
 
 	// list all geofences
-	listGeofences(
-		options?: ListGeofenceOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
-	): Promise<ListGeofenceResults>;
+	listGeofences(options?: ListGeofenceOptions): Promise<ListGeofenceResults>;
 
 	// Delete geofences
 	deleteGeofences(
 		geofenceIds: string[],
-		options?: GeofenceOptions,
-		customUserAgentDetails?: CustomUserAgentDetails
+		options?: GeofenceOptions
 	): Promise<DeleteGeofencesResults>;
 }
