+++ packages/predictions/package.json
@@ -69,7 +69,7 @@
 			"name": "Predictions (Convert provider)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, Predictions, AmazonAIConvertPredictionsProvider }",
-			"limit": "58 kB"
+			"limit": "57.9 kB"
 		},
 		{
 			"name": "Predictions (Identify provider)",
